requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
typer==0.12.5
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
rich==13.9.3
PyYAML==6.0.2
openai==1.52.1
